update&delete buttons if user is author, return HttpResponse if not

+++ src/blog/views.py
@@ -49,6 +49,8 @@ def post_detail(request, slug):
 def post_update(request, slug):
     obj = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)  # instance brings the form filled with the data
+    if request.user.id != obj.author.id:
+        return HttpResponse("You are not authorized!")
     if form.is_valid():
         form.save()
         return redirect("blog:list")
@@ -61,6 +63,8 @@ def post_update(request, slug):
     
 def post_delete(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id:
+        return HttpResponse("You are not authorized!")
     if request.method == "POST":
         obj.delete()
         return redirect("blog:list")
